refactoring to make pipeline.py simpler
- create get_train_val_datasets() function to get datasets for training
- move targets_multi_output() from model_setup.py to utils.py for more
  flexible access (solving import loop issue)
